Add Request_Constants module
