+++ common/config/request_constants.py
@@ -1,0 +1,34 @@
+'''
+Created on Nov 4, 2018
+
+@author: agagaleza
+'''
+class RequestConstants(object):
+    '''
+    
+    '''
+    #request json body
+    JSON_STATUS_AVAILABLE = 'available'
+    JSON_STATUS_PENDING = 'pending'
+    JSON_STATUS_SOLD = 'sold'
+    JSON_NAME = 'name'
+    JSON_PHOTOURLS = 'photoUrls'
+    JSON_STATUS = 'status'
+    JSON_UPLOAD_IMAGE = 'uploadImage'
+    JSON_ID = 'id'
+    JSON_MESSAGE = 'message'
+    JSON_GET = 'GET'
+    JSON_FINDBYSTATUS = 'GET FINDBYSTATUS'
+    JSON_DELETE = 'DELETE'
+    JSON_POST = 'POST'
+    JSON_PUT = 'PUT'
+    JSON_UPLOAD = 'POST UPLOADIMAGE'
+    JSON_REQUEST_STATUS = '?status='
+    JSON_URL = 'basic_url'
+    JSON_CONTENT_TYPE = 'Content-Type'
+    JSON_ACCEPT = 'accept'
+    JSON_ENDPOINT = 'api_endpoint'
+    JSON_RESPONSE = 'response_full'
+    JSON_RESPONSE_CODE = 'expected_response_code'
+    JSON_RESPONSE_CONTENT_TYPE =  'expected_response_content_type'
+
