+++ features/domain_models/pet.py
@@ -6,6 +6,7 @@ Created on Sep 15, 2018
 
 from common.util.random_string_generator import RandomStringGenerator
 import logging
+from common.config.request_constants import RequestConstants
 
 # -----------------------------------------------------------------------------
 # DOMAIN-MODEL:
@@ -15,7 +16,7 @@ class Pet(object):
     '''
     classdocs
     '''
-    status_list = ["available", "pending", "sold"]
+    status_list = [RequestConstants.JSON_STATUS_AVAILABLE, RequestConstants.JSON_STATUS_PENDING, RequestConstants.JSON_STATUS_SOLD]
     pet_id = 0
     name = ""
     photourls = []
@@ -51,10 +52,10 @@ class Pet(object):
         self.photo = photo
             
     def set_pet_details(self, pet_details):
-        self.set_pet_name(pet_details.get("name"))
+        self.set_pet_name(pet_details.get(RequestConstants.JSON_NAME))
 #         for pet_URLS
-        self.set_pet_photourls(pet_details.get("photoUrls"))
-        self.set_pet_status(pet_details.get("status"))
+        self.set_pet_photourls(pet_details.get(RequestConstants.JSON_PHOTOURLS))
+        self.set_pet_status(pet_details.get(RequestConstants.JSON_STATUS))
         
     def __init__(self):
         '''
